Refactor on SiteSwitcherSelect.tsx
