+++ ui/app/src/utils/registerComponents.ts
@@ -97,6 +97,7 @@ import PublishingStatusButton from '../components/PublishingStatusButton';
 import QuickCreate from '../modules/Preview/QuickCreate';
 import EditModeSwitch from '../components/EditModeSwitch';
 import AddressBar from '../components/PreviewAddressBar';
+import SiteSwitcherSelect from '../components/SiteSwitcherSelect';
 
 export const registerComponents = () => {
   Object.entries({
@@ -181,7 +182,8 @@ export const registerComponents = () => {
     'craftercms.components.PublishingStatusButton': PublishingStatusButton,
     'craftercms.components.QuickCreate': QuickCreate,
     'craftercms.components.EditModeSwitch': EditModeSwitch,
-    'craftercms.components.AddressBar': AddressBar
+    'craftercms.components.AddressBar': AddressBar,
+    'craftercms.components.SiteSwitcherSelect': SiteSwitcherSelect
   }).forEach(([id, component]) => {
     components.set(id, component);
   });
